pyproject.toml
[project]
name = "batch-classifier"
version = "0.1.0"
description = "Classify text data from a CSV file with the OpenAI batch API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "litellm>=1.57.5",
    "nest-asyncio>=1.6.0",
    "pandas>=2.2.3",
    "pytest-asyncio>=0.25.2",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = [
    "classifier",
    "process_batch",
    "prompt",
    "submit_batch"
]

[tool.pytest.ini_options]
pythonpath = "."
